requirements.txt
aio-pika
PyYAML
aiopg
aiohttp
git+git://github.com/NQN-Discord/rabbit_helper.git